# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Analytics Model API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /features/pregame/game/{gameId}:
    get:
      operationId: getPregameFeaturesForGame
      tags: [nfl]
      summary: Read versioned pregame features for one game
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Team-level pregame feature rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PregameFeatureRow"
  /features/audit:
    get:
      operationId: listPregameFeatureAudit
      tags: [dashboard]
      summary: Read feature values and point-in-time source audit metadata
      parameters:
        - name: season
          in: query
          required: false
          schema: { type: integer }
        - name: week
          in: query
          required: false
          schema: { type: integer }
        - name: gameId
          in: query
          required: false
          schema: { type: string }
        - name: teamId
          in: query
          required: false
          schema: { type: string }
        - name: featureVersion
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Filtered feature audit rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PregameFeatureAuditRow"
  /features/pregame/build:
    post:
      operationId: buildPregameFeatures
      tags: [dashboard]
      summary: Build an immutable version of historical pregame features
      description: Requires authenticated administrator access. The optional featureVersion pregame-v4-personnel-context creates additive rows from pregame-v3 and never overwrites v3. Never trains or scores a betting model.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PregameFeatureBuildInput"
      responses:
        "200":
          description: Feature build result
        "401":
          description: Authentication required
        "403":
          description: Administrator access required
  /features/personnel-context/game/{gameId}:
    get:
      operationId: getPersonnelContextForGame
      tags: [dashboard]
      summary: Read auditable point-in-time personnel and context intelligence
      description: Returns probable starters, QB certainty, injury impacts, OL continuity, unit matchups, explicit weather/schedule limitations, immutable market movement, and non-betting data confidence.
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Personnel and context audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonnelContext"
        "404":
          description: Game not found
  /features/personnel-context/audit:
    get:
      operationId: listPersonnelContextAudit
      tags: [dashboard]
      summary: List persisted Phase 7 personnel-context audit rows
      parameters:
        - name: gameId
          in: query
          schema: { type: string }
        - name: teamId
          in: query
          schema: { type: string }
        - name: season
          in: query
          schema: { type: integer }
        - name: week
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: Personnel-context audit rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonnelContextAuditRow"
...
components:
...
  schemas:
...
    PregameFeatureRow:
      type: object
      required: [featureVersion, gameId, teamId, opponentTeamId, season, week, kickoffTime, isHome, features, sampleCounts, featureAudit, lowSample, sourceCutoff]
      properties:
        featureVersion: { type: string }
        gameId: { type: string }
        teamId: { type: string }
        opponentTeamId: { type: string }
        season: { type: integer }
        week: { type: integer }
        kickoffTime: { type: string, format: date-time }
        isHome: { type: boolean }
        features: { type: object, additionalProperties: { type: ["number", "null"] } }
        sampleCounts: { type: object, additionalProperties: { type: integer } }
        featureAudit: { type: object, additionalProperties: true }
        lowSample: { type: boolean }
        sourceCutoff: { type: string, format: date-time }
        generatedAt: { type: string, format: date-time }
...
    PregameFeatureAuditRow:
      type: object
      required: [featureVersion, gameId, teamId, season, week, kickoffTime, featureName, value, sourceDataset, lookbackWindow, gamesIncluded, sampleSize, quality]
      properties:
        featureVersion: { type: string }
        gameId: { type: string }
        teamId: { type: string }
        opponentTeamId: { type: string }
        season: { type: integer }
        week: { type: integer }
        kickoffTime: { type: string, format: date-time }
        featureName: { type: string }
        value: { type: ["number", "null"] }
        sourceDataset: { type: string }
        lookbackWindow: { type: string }
        gamesIncluded: { type: integer }
        lastSourceGame: { type: ["string", "null"] }
        lastSourceDate: { type: ["string", "null"] }
        sampleSize: { type: integer }
        quality: { type: string, enum: [high, low_sample, unavailable] }
        unavailableReason: { type: ["string", "null"] }
    PersonnelContext:
      type: object
      required: [version, gameId, kickoffTime, sourceCutoff, teams, matchup, weather, market, dataConfidence, sources, limitations]
      properties:
        version: { type: string, enum: [pregame-v4-personnel-context] }
        gameId: { type: string }
        kickoffTime: { type: ["string", "null"], format: date-time }
        sourceCutoff: { type: string, format: date-time }
        teams: { type: object, additionalProperties: true }
        matchup: { type: array, items: { type: object, additionalProperties: true } }
        weather: { type: object, additionalProperties: true }
        market: { type: object, additionalProperties: true }
        dataConfidence: { type: object, additionalProperties: true }
        sources: { type: array, items: { type: string } }
        limitations: { type: array, items: { type: string } }
    PersonnelContextAuditRow:
      type: object
      required: [featureVersion, gameId, teamId, opponentTeamId, season, week, kickoffTime, sourceCutoff, dataConfidence, featureAudit, sources]
      properties:
        featureVersion: { type: string, enum: [pregame-v4-personnel-context] }
        gameId: { type: string }
        teamId: { type: string }
        opponentTeamId: { type: string }
        season: { type: integer }
        week: { type: integer }
        kickoffTime: { type: string, format: date-time }
        sourceCutoff: { type: string, format: date-time }
        dataConfidence: { type: ["number", "null"] }
        featureAudit: { type: object, additionalProperties: true }
        sources: { type: array, items: { type: string } }
...
    PregameFeatureBuildInput:
      type: object
      properties:
        featureVersion:
          type: string
          enum: [pregame-v3, pregame-v4-personnel-context]